--- src/typings/api_v1/user.d.ts
@@ -1,9 +1,0 @@
-// 创作类型声明
-export type userCreationData = {
-  pictureCount: number;
-  polygenCount: number;
-  videoCount: number;
-  postCount: number;
-  likeCount: number;
-  verseCount: number;
-};
